Making sodium resolver safer
