+++ exonum/src/crypto/mod.rs
@@ -69,6 +69,9 @@ use encoding::{serialize::{encode_hex, FromHex, FromHexError, ToHex},
                Offset};
 use helpers::Round;
 
+#[macro_use]
+mod macros;
+
 pub mod x25519;
 
 /// The size to crop the string in debug messages.
@@ -226,7 +229,7 @@ pub fn init() {
 /// }
 /// let _ = hash_stream.hash();
 /// ```
-#[derive(Debug, Default)]
+#[derive(Clone, Debug, Default)]
 pub struct HashStream(HashState);
 
 impl HashStream {
@@ -360,123 +363,6 @@ impl SignStream {
     }
 }
 
-macro_rules! implement_public_sodium_wrapper {
-    ($(#[$attr:meta])* struct $name:ident, $name_from:ident, $size:expr) => (
-    #[derive(PartialEq, Eq, Clone, Copy, PartialOrd, Ord, Hash)]
-    $(#[$attr])*
-    pub struct $name($name_from);
-
-    impl $name {
-        /// Creates a new instance filled with zeros.
-        pub fn zero() -> Self {
-            $name::new([0; $size])
-        }
-    }
-
-    impl $name {
-        /// Creates a new instance from bytes array.
-        pub fn new(bytes_array: [u8; $size]) -> Self {
-            $name($name_from(bytes_array))
-        }
-
-        /// Creates a new instance from bytes slice.
-        pub fn from_slice(bytes_slice: &[u8]) -> Option<Self> {
-            $name_from::from_slice(bytes_slice).map($name)
-        }
-
-        /// Returns a hex representation of binary data.
-        /// Lower case letters are used (e.g. f9b4ca).
-        pub fn to_hex(&self) -> String {
-            encode_hex(self)
-        }
-    }
-
-    impl AsRef<[u8]> for $name {
-        fn as_ref(&self) -> &[u8] {
-            self.0.as_ref()
-        }
-    }
-
-    impl FromStr for $name {
-        type Err = FromHexError;
-        fn from_str(s: &str) -> Result<Self, Self::Err> {
-            $name::from_hex(s)
-        }
-    }
-
-    impl fmt::Debug for $name {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, stringify!($name))?;
-            write!(f, "(")?;
-            for i in &self[0..BYTES_IN_DEBUG] {
-                write!(f, "{:02X}", i)?
-            }
-            write!(f, ")")
-        }
-    }
-
-    impl fmt::Display for $name {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            f.write_str(&self.to_hex())
-        }
-    }
-    )
-}
-
-macro_rules! implement_private_sodium_wrapper {
-    ($(#[$attr:meta])* struct $name:ident, $name_from:ident, $size:expr) => (
-    #[derive(Clone, PartialEq, Eq)]
-    $(#[$attr])*
-    pub struct $name($name_from);
-
-    impl $name {
-        /// Creates a new instance filled with zeros.
-        pub fn zero() -> Self {
-            $name::new([0; $size])
-        }
-    }
-
-    impl $name {
-        /// Creates a new instance from bytes array.
-        pub fn new(bytes_array: [u8; $size]) -> Self {
-            $name($name_from(bytes_array))
-        }
-
-        /// Creates a new instance from bytes slice.
-        pub fn from_slice(bytes_slice: &[u8]) -> Option<Self> {
-            $name_from::from_slice(bytes_slice).map($name)
-        }
-
-        /// Returns a hex representation of binary data.
-        /// Lower case letters are used (e.g. f9b4ca).
-        pub fn to_hex(&self) -> String {
-            encode_hex(&self[..])
-        }
-    }
-
-    impl fmt::Debug for $name {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, stringify!($name))?;
-            write!(f, "(")?;
-            for i in &self[0..BYTES_IN_DEBUG] {
-                write!(f, "{:02X}", i)?
-            }
-            write!(f, "...)")
-        }
-    }
-
-    impl ToHex for $name {
-        fn write_hex<W: ::std::fmt::Write>(&self, w: &mut W) -> ::std::fmt::Result {
-            (self.0).0.as_ref().write_hex(w)
-        }
-
-        fn write_hex_upper<W: ::std::fmt::Write>(&self, w: &mut W) -> ::std::fmt::Result {
-            (self.0).0.as_ref().write_hex_upper(w)
-        }
-    }
-    )
-}
-
 implement_public_sodium_wrapper! {
 /// Ed25519 public key used to verify digital signatures.
 ///
@@ -599,94 +485,12 @@ implement_private_sodium_wrapper! {
     struct Seed, SeedSodium, SEED_LENGTH
 }
 
-macro_rules! implement_serde {
-    ($name:ident) => {
-        impl FromHex for $name {
-            type Error = FromHexError;
-
-            fn from_hex<T: AsRef<[u8]>>(v: T) -> Result<Self, Self::Error> {
-                let bytes = Vec::<u8>::from_hex(v)?;
-                if let Some(self_value) = Self::from_slice(bytes.as_ref()) {
-                    Ok(self_value)
-                } else {
-                    Err(FromHexError::InvalidStringLength)
-                }
-            }
-        }
-
-        impl Serialize for $name {
-            fn serialize<S>(&self, ser: S) -> Result<S::Ok, S::Error>
-            where
-                S: Serializer,
-            {
-                let hex_string = encode_hex(&self[..]);
-                ser.serialize_str(&hex_string)
-            }
-        }
-
-        impl<'de> Deserialize<'de> for $name {
-            fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-            where
-                D: Deserializer<'de>,
-            {
-                struct HexVisitor;
-
-                impl<'v> Visitor<'v> for HexVisitor {
-                    type Value = $name;
-                    fn expecting(&self, fmt: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-                        write!(fmt, "expecting str.")
-                    }
-                    fn visit_str<E>(self, s: &str) -> Result<Self::Value, E>
-                    where
-                        E: de::Error,
-                    {
-                        $name::from_hex(s).map_err(|_| de::Error::custom("Invalid hex"))
-                    }
-                }
-                deserializer.deserialize_str(HexVisitor)
-            }
-        }
-    };
-}
-
 implement_serde! {Hash}
 implement_serde! {PublicKey}
 implement_serde! {SecretKey}
 implement_serde! {Seed}
 implement_serde! {Signature}
 
-macro_rules! implement_index_traits {
-    ($new_type:ident) => {
-        impl Index<Range<usize>> for $new_type {
-            type Output = [u8];
-            fn index(&self, _index: Range<usize>) -> &[u8] {
-                let inner = &self.0;
-                inner.0.index(_index)
-            }
-        }
-        impl Index<RangeTo<usize>> for $new_type {
-            type Output = [u8];
-            fn index(&self, _index: RangeTo<usize>) -> &[u8] {
-                let inner = &self.0;
-                inner.0.index(_index)
-            }
-        }
-        impl Index<RangeFrom<usize>> for $new_type {
-            type Output = [u8];
-            fn index(&self, _index: RangeFrom<usize>) -> &[u8] {
-                let inner = &self.0;
-                inner.0.index(_index)
-            }
-        }
-        impl Index<RangeFull> for $new_type {
-            type Output = [u8];
-            fn index(&self, _index: RangeFull) -> &[u8] {
-                let inner = &self.0;
-                inner.0.index(_index)
-            }
-        }
-    };
-}
 implement_index_traits! {Hash}
 implement_index_traits! {PublicKey}
 implement_index_traits! {SecretKey}
